Move JobGroup Cancelling to After Failure DelayedJob Lifecycle Hook
Currently, the `on_cancellation_job` will trigger in the before failure
hook of a job. This essentially makes it impossible for a job to perform
any kind of cleanup on failure before the cancellation job is triggered.

This can lead to a race condition where other failure hooks are running
at the same time as the cancellation job.

This change proposes to move the cancellation to the after failure hook,
allowing jobs to define cleanup logic in the before hook, before the
cancellation job is triggered.