Move On Cancellation Logic to After Failure Hook
In order to allow individual jobs in the JobGroup to perform cleanup activities upon 
job failure, a JobGroups on cancellation job must be enqueued in the after failure 
delayed job lifecycle hook. If both were to be performed in the same hook, the job's
failure hook and the on cancellation job may run at the same time since hook execution 
order is never guaranteed. This is particular important if the on cancellation job 
depends on state set in the failure hook of an individual job.

This also forces the Delayed::Worker.destroy_failed_jobs to be set to
false in order to prevent a race condition where a JobGroup could be
completed (instead of cancelled) if a successful job occurs at the same
time as a failing job.